+++ src/crystal/model/resource.py
@@ -17,8 +17,9 @@ from crystal.util.db import (
 )
 from crystal.util.profile import warn_if_slow
 from crystal.util.urls import requote_uri
+from crystal.util.xfutures import warn_if_result_not_read
 from crystal.util.xthreading import (
-    fg_affinity, is_foreground_thread,
+    fg_affinity, is_foreground_thread, scheduler_affinity
 )
 import itertools
 from typing import (
@@ -1048,10 +1049,31 @@ class Resource:
         return True
     
     # NOTE: Only used from a Python REPL at the moment
-    def delete(self) -> None:
+    def delete(self) -> Future[None]:
         """
-        Deletes this resource, including any related revisions.
-        
+        Deletes this resource asynchronously, including any related revisions.
+        Returns a Future that resolves when deletion is complete.
+
+        Future Raises:
+        * ProjectReadOnlyError
+        * ValueError -- if this resource is referenced by a RootResource
+        * sqlite3.DatabaseError, OSError --
+            if the delete partially/fully failed, leaving behind zero or more revisions
+        """
+        from crystal.task import DeleteResourceTask
+        task = DeleteResourceTask(self)
+        self.project.add_task(task)
+        # NOTE: Crystal <=2.2.0 did NOT return a Future, so older callers may
+        #       not expect a Future and in particular may not wait for it properly.
+        #       Print a warning to make those now-improper callers easier to identify.
+        return warn_if_result_not_read(task.future, f'{self}.delete()')
+
+    @fg_affinity  # does database I/O
+    @scheduler_affinity  # does revision body I/O
+    def _delete_now(self) -> None:
+        """
+        Deletes this resource synchronously, including any related revisions.
+
         Raises:
         * ProjectReadOnlyError
         * ValueError -- if this resource is referenced by a RootResource
@@ -1069,7 +1091,7 @@ class Resource:
         with closing(project._db.cursor()) as c:
             root_resource_ids = [
                 id
-                for (id,) in 
+                for (id,) in
                 c.execute('select id from root_resource where resource_id=?', (self._id,))
             ]
         if len(root_resource_ids) > 0:
@@ -1081,7 +1103,7 @@ class Resource:
         #       including the default_revision(), will be left intact, which
         #       is what most users of Resource care about.
         for rev in list(self.revisions()):
-            rev.delete()
+            rev._delete_now()
         
         project._resource_will_delete(self)